AddFunctions & ChangeToArray
An array variable called cart was created to store the order of the user. Also the object stored on the variable named "cart", change the type to array and the variable name to "store". The objects inside it took on a new property called "stock", which stores the amount of merchandise available.
Some functions were created to work with these new variables.